refactored calls to previous logger
