+++ paos/__init__.py
@@ -31,18 +31,8 @@ if base_dir is not None and os.path.exists(os.path.join(base_dir, ".git")):
 else:
     __commit__ = None
 
-# initialise logger
-import logging
-
-logger = logging.getLogger(__pkg_name__)
-logger.info(f"code version {__version__}")
-
-from paos.log import setLogLevel
-
-setLogLevel(logging.INFO)
-
-# from paos.log import addLogFile
-# addLogFile(level=logging.INFO)
+from loguru import logger
+logger.level("Announce", no=100, color="<magenta>")
 
 from paos.classes.wfo import WFO
 from paos.classes.abcd import ABCD
